move tune_analysis.constants to definitions.constants
